rename nwi-map-controls-v2 to nwi-map-search and remove unused component
