+++ src/app/views/river-index/components/index.js
@@ -5,6 +5,6 @@ export { default as NwiFullscreenToggle } from './nwi-fullscreen-toggle.vue'
 export { default as NwiBasemapToggle } from './nwi-basemap-toggle.vue'
 export { default as NwiColorByToggle } from './nwi-color-by-toggle.vue'
 export { default as NwiMapControls } from './nwi-map-controls.vue'
-export { default as NwiMapControlsV2 } from './nwi-map-controls-v2.vue'
+export { default as NwiMapSearch } from './nwi-map-search.vue'
 export { default as NwiRiversTable } from './nwi-rivers-table.vue'
 export { default as NwiMap } from './nwi-map.vue'
