+++ src/app/views/river-index/river-index.vue
@@ -21,7 +21,7 @@
           />
         </div>
         <div class="bx--col-sm-16 bx--col-lg-6">
-          <NwiMapControlsV2 />
+          <NwiMapSearch />
           <template v-if="loading">
             <UtilityBlock state="loading" />
           </template>
@@ -38,7 +38,7 @@
 </template>
 
 <script>
-import { NwiRiversTable, NwiMap, NwiMapControlsV2 } from './components'
+import { NwiRiversTable, NwiMap, NwiMapSearch } from './components'
 import 'mapbox-gl/dist/mapbox-gl.css'
 import { riverIndexActions } from './shared/state'
 import { mapState } from 'vuex'
@@ -54,7 +54,7 @@ export default {
     NwiMap,
     NwiRiversTable,
     UtilityBlock,
-    NwiMapControlsV2
+    NwiMapSearch
   },
   mixins: [riverSearchHttpConfig],
   props: {
